LOJ-1009 (Back to Underworld) Editorial in English (#207)
* 1212-LOJ English tutorial added

Added the English tutorial per community guidelines for this problem.

* small fix in the editorial

* Added the recommended space after headers '#'

There was a small issue where I forgot to put a space after '#' signs making it appear in the markdown. This fixes that issue. Also changed the "cplusplus" to "cpp" in the blockquote so code appears properly highlighted.

* LOJ-1009 (Back to Underworld) editorial added

Added the English tutorial for LOJ-1009  (Back to Underworld) . Please check the updates.

* Added the needed space .

There was some issue where the lack of  space showed some markdown as typos , particularly not bolding some parts of the text. hopefully that has been resolved.